requirements.txt
# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# OpenAI
openai>=1.0.0

# HTTP requests
requests>=2.31.0

# Configuration
PyYAML>=6.0.1

# Email processing
email-validator>=2.0.0

# Logging and utilities
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
